requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
numpy>=1.20.0
matplotlib>=3.5.0
tensorboard>=2.12.0
pyyaml>=6.0.0
scikit-learn>=1.6.0
seaborn>=0.13.0
pandas>=2.0.0

# Optional monitoring enhancements
gputil>=1.4.0; platform_system != "Windows"  # GPU monitoring (Linux/Mac only)
psutil>=5.9.0  # System resource monitoring
torchviz>=0.0.2  # Enhanced model visualization
graphviz>=0.20.1  # Required for torchviz